# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EVENTS - тестовый проект",
        "description": "API управления Событиями проекта",
        "contact": {
            "name": "Stanislav",
            "email": "postnikov.sa@ya.ru"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Пользователи"
                ],
                "summary": "Регистрация нового пользователя",
                "description": "Регистрирует нового пользователя с предоставленным логином, email, паролем, и датой рождения (необязательно), а также именем и фамилией.",
                "operationId": "f6aac423803912ae30170e84617813e9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "email",
                                    "password",
                                    "first_name",
                                    "last_name"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "Ivan123"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "Ivan123@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01",
                                        "nullable": true
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "Ivan"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Petrov"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пользователь успешно зарегистрирован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "result": {
                                            "$ref": "/docs/swagger.yaml#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "The login has already been taken."
                                                ]
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Переданы некорректные данные."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Пользователи"
                ],
                "summary": "Авторизация пользователя",
                "description": "Авторизует пользователя на основе предоставленных учетных данных, и выдает токен для дальнейшего взаимодействия",
                "operationId": "e83d17e49cd199471719b81efec9ed14",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "Ivan123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная авторизация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "result": {
                                            "type": "string",
                                            "example": "Успешная авторизация"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неправильные учетные данные",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Неправильные учетные данные"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}